fixed issues with maping table with TS
