+++ package.json
@@ -17,6 +17,7 @@
     "gh-pages": "^3.2.3",
     "joi": "^17.6.0",
     "joi-browser": "13.4",
+    "lodash": "^4.17.21",
     "query-string": "^6.1.0",
     "react": "^17.0.2",
     "react-big-calendar": "^0.38.2",
